Put types in shared
