+++ packages/shared/index.mts
@@ -5,6 +5,7 @@ export * from './src/schemas/websockets.mjs';
 export * from './src/types/apps.mjs';
 export * from './src/types/cells.mjs';
 export * from './src/types/tsserver.mjs';
+export * from './src/types/history.mjs';
 export * from './src/types/websockets.mjs';
 export * from './src/types/secrets.mjs';
 export * from './src/utils.mjs';
